update code server(target) kdw001

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/log/KrcmtEmpExeTarget.java
@@ -30,10 +30,10 @@ public class KrcmtEmpExeTarget extends UkJpaEntity implements Serializable  {
 	public KrcmtEmpExeTargetPK krcmtEmpExeTargetPK;
 	
 	@Column(name = "EXECUTION_CONTENT")
-	public BigDecimal executionContent;
+	public int executionContent;
 
 	@Column(name = "EXECUTION_STATE")
-	public BigDecimal executionState;
+	public int executionState;
 
 	@Override
 	protected Object getKey() {
